Update dbcc-show-statistics-transact-sql.md
in the case of a step with only rows with the upper bound's value, you have range_rows 0 and  distinct_range_rows 0 (since these two exlude the upper bound), but the AVG_RANGE_ROWS is 1, this result leads us to understand that the upper bound is not considered when there are different values other than the upper bound value. but when there is only one single value (the upper bound's one) this distinct value is considered.  0/0 cannot give 1 :)